refactor: add the separate args module
